agents/code_interpreter/requirements.txt
langgraph>=0.2.0
python-dotenv>=1.0.1
openai>=1.40.0
google-genai>=0.5.0
